requirements.txt
django
djangorestframework
djangorestframework-jwt
django-filter
drf_yasg
django-cors-headers